requirements.txt
# ── Core AI & LLM ──────────────────────────
openai>=1.0.0
anthropic>=0.20.0
groq>=0.4.0
google-generativeai>=0.5.0

# ── Environment & Config ────────────────────
python-dotenv>=1.0.0

# ── HTTP Requests ───────────────────────────
requests>=2.31.0
httpx>=0.25.0

# ── Data Handling ───────────────────────────
pandas>=2.0.0
numpy>=1.24.0

# ── Web Scraping ────────────────────────────
beautifulsoup4>=4.12.0
selenium>=4.15.0

# ── Logging & Utilities ─────────────────────
loguru>=0.7.0
rich>=13.0.0
tqdm>=4.66.0

# ── Testing ─────────────────────────────────
pytest>=7.4.0